Separated socket related code from app.js (formerly main.js)
