+++ bootstrapper/app/router/socket.js
@@ -1,0 +1,9 @@
+'use strict'
+
+module.exports = socket => {
+
+  socket.on('client:perform', request => {
+    console.log(request);
+  });
+
+};
